requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
mysql-connector-python==8.0.33